Move fetch into separate service, add error handling, don't log everything for some objects.
