+++ apps/server/src/infra/sync/tsp/tsp-sync.service.spec.ts
@@ -1,36 +1,23 @@
 import { faker } from '@faker-js/faker';
 import { createMock, DeepMocked } from '@golevelup/ts-jest';
-import {
-	ExportApiInterface,
-	RobjExportKlasse,
-	RobjExportLehrer,
-	RobjExportLehrerMigration,
-	RobjExportSchueler,
-	RobjExportSchuelerMigration,
-	RobjExportSchule,
-	TspClientFactory,
-} from '@infra/tsp-client';
 import { School, SchoolService } from '@modules/school';
 import { SystemService, SystemType } from '@modules/system';
 import { Test, TestingModule } from '@nestjs/testing';
 import { SystemProvisioningStrategy } from '@shared/domain/interface/system-provisioning.strategy';
 import { federalStateFactory, schoolYearFactory, userDoFactory } from '@shared/testing';
-import { AccountService } from '@src/modules/account';
+import { AccountService } from '@modules/account';
 import { accountDoFactory } from '@src/modules/account/testing';
-import { FederalStateService, SchoolYearService } from '@src/modules/legacy-school';
-import { OauthConfigMissingLoggableException } from '@src/modules/oauth/loggable';
+import { FederalStateService, SchoolYearService } from '@modules/legacy-school';
 import { SchoolProps } from '@src/modules/school/domain';
 import { FederalStateEntityMapper, SchoolYearEntityMapper } from '@src/modules/school/repo/mikro-orm/mapper';
 import { schoolFactory } from '@src/modules/school/testing';
 import { systemFactory } from '@src/modules/system/testing';
-import { UserService } from '@src/modules/user';
-import { AxiosResponse } from 'axios';
+import { UserService } from '@modules/user';
 import { TspSyncService } from './tsp-sync.service';
 
 describe(TspSyncService.name, () => {
 	let module: TestingModule;
 	let sut: TspSyncService;
-	let tspClientFactory: DeepMocked<TspClientFactory>;
 	let systemService: DeepMocked<SystemService>;
 	let schoolService: DeepMocked<SchoolService>;
 	let federalStateService: DeepMocked<FederalStateService>;
@@ -42,10 +29,6 @@ describe(TspSyncService.name, () => {
 		module = await Test.createTestingModule({
 			providers: [
 				TspSyncService,
-				{
-					provide: TspClientFactory,
-					useValue: createMock<TspClientFactory>(),
-				},
 				{
 					provide: SystemService,
 					useValue: createMock<SystemService>(),
@@ -74,7 +57,6 @@ describe(TspSyncService.name, () => {
 		}).compile();
 
 		sut = module.get(TspSyncService);
-		tspClientFactory = module.get(TspClientFactory);
 		systemService = module.get(SystemService);
 		schoolService = module.get(SchoolService);
 		federalStateService = module.get(FederalStateService);
@@ -127,235 +109,6 @@ describe(TspSyncService.name, () => {
 				setup();
 
 				await expect(sut.findTspSystemOrFail()).rejects.toThrow();
-			});
-		});
-	});
-
-	const setupTspClient = () => {
-		const clientId = faker.string.alpha();
-		const clientSecret = faker.string.alpha();
-		const tokenEndpoint = faker.internet.url();
-		const system = systemFactory.build({
-			oauthConfig: {
-				clientId,
-				clientSecret,
-				tokenEndpoint,
-			},
-		});
-
-		const tspSchool: RobjExportSchule = {
-			schuleName: faker.string.alpha(),
-			schuleNummer: faker.string.alpha(),
-		};
-		const schools = [tspSchool];
-		const responseSchools = createMock<AxiosResponse<Array<RobjExportSchule>>>({
-			data: schools,
-		});
-
-		const tspTeacher: RobjExportLehrer = {
-			schuleNummer: faker.string.alpha(),
-			lehrerVorname: faker.string.alpha(),
-			lehrerNachname: faker.string.alpha(),
-			lehrerUid: faker.string.alpha(),
-		};
-		const teachers = [tspTeacher];
-		const responseTeachers = createMock<AxiosResponse<Array<RobjExportLehrer>>>({
-			data: teachers,
-		});
-
-		const tspStudent: RobjExportSchueler = {
-			schuleNummer: faker.string.alpha(),
-			schuelerVorname: faker.string.alpha(),
-			schuelerNachname: faker.string.alpha(),
-			schuelerUid: faker.string.alpha(),
-		};
-		const students = [tspStudent];
-		const responseStudents = createMock<AxiosResponse<Array<RobjExportSchueler>>>({
-			data: students,
-		});
-
-		const tspClass: RobjExportKlasse = {
-			schuleNummer: faker.string.alpha(),
-			klasseId: faker.string.alpha(),
-			klasseName: faker.string.alpha(),
-			lehrerUid: faker.string.alpha(),
-		};
-		const classes = [tspClass];
-		const responseClasses = createMock<AxiosResponse<Array<RobjExportKlasse>>>({
-			data: classes,
-		});
-
-		const tspTeacherMigration: RobjExportLehrerMigration = {
-			lehrerUidAlt: faker.string.alpha(),
-			lehrerUidNeu: faker.string.alpha(),
-		};
-		const teacherMigrations = [tspTeacherMigration];
-		const responseTeacherMigrations = createMock<AxiosResponse<Array<RobjExportLehrerMigration>>>({
-			data: teacherMigrations,
-		});
-
-		const tspStudentMigration: RobjExportSchuelerMigration = {
-			schuelerUidAlt: faker.string.alpha(),
-			schuelerUidNeu: faker.string.alpha(),
-		};
-		const studentMigrations = [tspStudentMigration];
-		const responseStudentMigrations = createMock<AxiosResponse<Array<RobjExportSchuelerMigration>>>({
-			data: studentMigrations,
-		});
-
-		const exportApiMock = createMock<ExportApiInterface>();
-		exportApiMock.exportSchuleList.mockResolvedValueOnce(responseSchools);
-		exportApiMock.exportLehrerList.mockResolvedValueOnce(responseTeachers);
-		exportApiMock.exportSchuelerList.mockResolvedValueOnce(responseStudents);
-		exportApiMock.exportKlasseList.mockResolvedValueOnce(responseClasses);
-		exportApiMock.exportLehrerListMigration.mockResolvedValueOnce(responseTeacherMigrations);
-		exportApiMock.exportSchuelerListMigration.mockResolvedValueOnce(responseStudentMigrations);
-
-		tspClientFactory.createExportClient.mockReturnValueOnce(exportApiMock);
-
-		return {
-			clientId,
-			clientSecret,
-			tokenEndpoint,
-			system,
-			exportApiMock,
-			schools,
-			teachers,
-			students,
-			classes,
-			teacherMigrations,
-			studentMigrations,
-		};
-	};
-
-	describe('fetchTspSchools', () => {
-		describe('when tsp schools are fetched', () => {
-			it('should use the oauthConfig to create the client', async () => {
-				const { clientId, clientSecret, tokenEndpoint, system } = setupTspClient();
-
-				await sut.fetchTspSchools(system, 1);
-
-				expect(tspClientFactory.createExportClient).toHaveBeenCalledWith({
-					clientId,
-					clientSecret,
-					tokenEndpoint,
-				});
-			});
-
-			it('should call exportSchuleList', async () => {
-				const { system, exportApiMock } = setupTspClient();
-
-				await sut.fetchTspSchools(system, 1);
-
-				expect(exportApiMock.exportSchuleList).toHaveBeenCalledTimes(1);
-			});
-
-			it('should return an array of schools', async () => {
-				const { system } = setupTspClient();
-
-				const schools = await sut.fetchTspSchools(system, 1);
-
-				expect(schools).toBeDefined();
-				expect(schools).toBeInstanceOf(Array);
-			});
-		});
-	});
-
-	describe('fetchTspTeachers', () => {
-		describe('when tsp teachers are fetched', () => {
-			it('should use the oauthConfig to create the client', async () => {
-				const { clientId, clientSecret, tokenEndpoint, system } = setupTspClient();
-
-				await sut.fetchTspTeachers(system, 1);
-
-				expect(tspClientFactory.createExportClient).toHaveBeenCalledWith({
-					clientId,
-					clientSecret,
-					tokenEndpoint,
-				});
-			});
-
-			it('should call exportLehrerList', async () => {
-				const { system, exportApiMock } = setupTspClient();
-
-				await sut.fetchTspTeachers(system, 1);
-
-				expect(exportApiMock.exportLehrerList).toHaveBeenCalledTimes(1);
-			});
-
-			it('should return an array of teachers', async () => {
-				const { system } = setupTspClient();
-
-				const teachers = await sut.fetchTspTeachers(system, 1);
-
-				expect(teachers).toBeDefined();
-				expect(teachers).toBeInstanceOf(Array);
-			});
-		});
-	});
-
-	describe('fetchTspStudents', () => {
-		describe('when tsp students are fetched', () => {
-			it('should use the oauthConfig to create the client', async () => {
-				const { clientId, clientSecret, tokenEndpoint, system } = setupTspClient();
-
-				await sut.fetchTspStudents(system, 1);
-
-				expect(tspClientFactory.createExportClient).toHaveBeenCalledWith({
-					clientId,
-					clientSecret,
-					tokenEndpoint,
-				});
-			});
-
-			it('should call exportSchuelerList', async () => {
-				const { system, exportApiMock } = setupTspClient();
-
-				await sut.fetchTspStudents(system, 1);
-
-				expect(exportApiMock.exportSchuelerList).toHaveBeenCalledTimes(1);
-			});
-
-			it('should return an array of students', async () => {
-				const { system } = setupTspClient();
-
-				const students = await sut.fetchTspStudents(system, 1);
-
-				expect(students).toBeDefined();
-				expect(students).toBeInstanceOf(Array);
-			});
-		});
-	});
-
-	describe('fetchTspClasses', () => {
-		describe('when tsp classes are fetched', () => {
-			it('should use the oauthConfig to create the client', async () => {
-				const { clientId, clientSecret, tokenEndpoint, system } = setupTspClient();
-
-				await sut.fetchTspClasses(system, 1);
-
-				expect(tspClientFactory.createExportClient).toHaveBeenCalledWith({
-					clientId,
-					clientSecret,
-					tokenEndpoint,
-				});
-			});
-
-			it('should call exportKlasseList', async () => {
-				const { system, exportApiMock } = setupTspClient();
-
-				await sut.fetchTspClasses(system, 1);
-
-				expect(exportApiMock.exportKlasseList).toHaveBeenCalledTimes(1);
-			});
-
-			it('should return an array of classes', async () => {
-				const { system } = setupTspClient();
-
-				const classes = await sut.fetchTspClasses(system, 1);
-
-				expect(classes).toBeDefined();
-				expect(classes).toBeInstanceOf(Array);
 			});
 		});
 	});
@@ -545,72 +298,6 @@ describe(TspSyncService.name, () => {
 		});
 	});
 
-	describe('fetchTspTeacherMigrations', () => {
-		describe('when tsp teacher migrations are fetched', () => {
-			it('should create the client', async () => {
-				const { clientId, clientSecret, tokenEndpoint, system } = setupTspClient();
-
-				await sut.fetchTspTeacherMigrations(system);
-
-				expect(tspClientFactory.createExportClient).toHaveBeenCalledWith({
-					clientId,
-					clientSecret,
-					tokenEndpoint,
-				});
-			});
-
-			it('should call exportLehrerListMigration', async () => {
-				const { system, exportApiMock } = setupTspClient();
-
-				await sut.fetchTspTeacherMigrations(system);
-
-				expect(exportApiMock.exportLehrerListMigration).toHaveBeenCalledTimes(1);
-			});
-
-			it('should return an array of teacher migrations', async () => {
-				const { system } = setupTspClient();
-
-				const result = await sut.fetchTspTeacherMigrations(system);
-
-				expect(result).toBeDefined();
-				expect(result).toBeInstanceOf(Array);
-			});
-		});
-	});
-
-	describe('fetchTspStudentMigrations', () => {
-		describe('when tsp student migrations are fetched', () => {
-			it('should create the client', async () => {
-				const { clientId, clientSecret, tokenEndpoint, system } = setupTspClient();
-
-				await sut.fetchTspStudentMigrations(system);
-
-				expect(tspClientFactory.createExportClient).toHaveBeenCalledWith({
-					clientId,
-					clientSecret,
-					tokenEndpoint,
-				});
-			});
-
-			it('should call exportSchuelerListMigration', async () => {
-				const { system, exportApiMock } = setupTspClient();
-
-				await sut.fetchTspStudentMigrations(system);
-
-				expect(exportApiMock.exportSchuelerListMigration).toHaveBeenCalledTimes(1);
-			});
-
-			it('should return an array of student migrations', async () => {
-				const { system } = setupTspClient();
-
-				const result = await sut.fetchTspStudentMigrations(system);
-
-				expect(result).toBeDefined();
-				expect(result).toBeInstanceOf(Array);
-			});
-		});
-	});
-
 	describe('findUserByTspUid', () => {
 		describe('when user is found', () => {
 			const setup = () => {
@@ -737,22 +424,6 @@ describe(TspSyncService.name, () => {
 				const result = await sut.updateAccount(account, username, systemId);
 
 				expect(result).toBe(account);
-			});
-		});
-	});
-
-	describe('createClient', () => {
-		describe('when oauthConfig is missing', () => {
-			const setup = () => {
-				const system = systemFactory.build({ oauthConfig: undefined });
-
-				return { system };
-			};
-
-			it('should throw an OauthConfigMissingLoggableException', async () => {
-				const { system } = setup();
-
-				await expect(async () => sut.fetchTspSchools(system, 1)).rejects.toThrow(OauthConfigMissingLoggableException);
 			});
 		});
 	});
