+++ apps/server/src/infra/sync/tsp/tsp-sync.service.ts
@@ -1,4 +1,3 @@
-import { TspClientFactory } from '@infra/tsp-client';
 import { FederalStateService, SchoolYearService } from '@modules/legacy-school';
 import { School, SchoolService } from '@modules/school';
 import { System, SystemService, SystemType } from '@modules/system';
@@ -8,13 +7,11 @@ import { SystemProvisioningStrategy } from '@shared/domain/interface/system-prov
 import { SchoolFeature } from '@shared/domain/types';
 import { Account, AccountService } from '@src/modules/account';
 import { FederalStateNames } from '@src/modules/legacy-school/types';
-import { OauthConfigMissingLoggableException } from '@src/modules/oauth/loggable';
 import { FederalState } from '@src/modules/school/domain';
 import { SchoolFactory } from '@src/modules/school/domain/factory';
 import { FederalStateEntityMapper, SchoolYearEntityMapper } from '@src/modules/school/repo/mikro-orm/mapper';
 import { UserService } from '@src/modules/user';
 import { ObjectId } from 'bson';
-import moment from 'moment/moment';
 import { TspSystemNotFoundLoggableException } from './loggable/tsp-system-not-found.loggable-exception';
 
 @Injectable()
@@ -22,7 +19,6 @@ export class TspSyncService {
 	private federalState: FederalState | undefined;
 
 	constructor(
-		private readonly tspClientFactory: TspClientFactory,
 		private readonly systemService: SystemService,
 		private readonly schoolService: SchoolService,
 		private readonly federalStateService: FederalStateService,
@@ -43,46 +39,6 @@ export class TspSyncService {
 		}
 
 		return systems[0];
-	}
-
-	public async fetchTspSchools(system: System, daysToFetch: number) {
-		const client = this.createClient(system);
-
-		const lastChangeDate = this.formatChangeDate(daysToFetch);
-		const schoolsResponse = await client.exportSchuleList(lastChangeDate);
-		const schools = schoolsResponse.data;
-
-		return schools;
-	}
-
-	public async fetchTspTeachers(system: System, daysToFetch: number) {
-		const client = this.createClient(system);
-
-		const lastChangeDate = this.formatChangeDate(daysToFetch);
-		const teachersResponse = await client.exportLehrerList(lastChangeDate);
-		const teachers = teachersResponse.data;
-
-		return teachers;
-	}
-
-	public async fetchTspStudents(system: System, daysToFetch: number) {
-		const client = this.createClient(system);
-
-		const lastChangeDate = this.formatChangeDate(daysToFetch);
-		const studentsResponse = await client.exportSchuelerList(lastChangeDate);
-		const students = studentsResponse.data;
-
-		return students;
-	}
-
-	public async fetchTspClasses(system: System, daysToFetch: number) {
-		const client = this.createClient(system);
-
-		const lastChangeDate = this.formatChangeDate(daysToFetch);
-		const classesResponse = await client.exportKlasseList(lastChangeDate);
-		const classes = classesResponse.data;
-
-		return classes;
 	}
 
 	public async findSchool(system: System, identifier: string): Promise<School | undefined> {
@@ -149,24 +105,6 @@ export class TspSyncService {
 		return this.federalState;
 	}
 
-	public async fetchTspTeacherMigrations(system: System) {
-		const client = this.createClient(system);
-
-		const teacherMigrationsResponse = await client.exportLehrerListMigration();
-		const teacherMigrations = teacherMigrationsResponse.data;
-
-		return teacherMigrations;
-	}
-
-	public async fetchTspStudentMigrations(system: System) {
-		const client = this.createClient(system);
-
-		const studentMigrationsResponse = await client.exportSchuelerListMigration();
-		const studentMigrations = studentMigrationsResponse.data;
-
-		return studentMigrations;
-	}
-
 	public async findUserByTspUid(tspUid: string): Promise<UserDO | null> {
 		const tspUser = await this.userService.findUsers({ tspUid });
 
@@ -207,23 +145,5 @@ export class TspSyncService {
 		account.systemId = systemId;
 
 		return this.accountService.save(account);
-	}
-
-	private formatChangeDate(daysToFetch: number): string {
-		return moment(new Date()).subtract(daysToFetch, 'days').subtract(1, 'hours').format('YYYY-MM-DD HH:mm:ss.SSS');
-	}
-
-	private createClient(system: System) {
-		if (!system.oauthConfig) {
-			throw new OauthConfigMissingLoggableException(system.id);
-		}
-
-		const client = this.tspClientFactory.createExportClient({
-			clientId: system.oauthConfig.clientId,
-			clientSecret: system.oauthConfig.clientSecret,
-			tokenEndpoint: system.oauthConfig.tokenEndpoint,
-		});
-
-		return client;
 	}
 }
